@@ -42,22 +42,20 @@
     "postcss-loader": "3.0.0",
     "postcss-preset-env": "6.0.6",
     "postcss-safe-parser": "4.0.1",
-    "react": "^16.4.1",
+    "react": "^16.8.6",
     "react-app-polyfill": "^0.1.3",
     "react-dev-utils": "^6.0.3",
-    "react-dom": "^16.4.1",
+    "react-dom": "^16.8.6",
     "react-responsive-modal": "^3.6.0",
     "react-reveal": "^1.2.2",
     "react-router-dom": "^4.3.1",
     "react-router-scroll-top": "^0.1.1",
-    "react-scroll-to-component": "^1.0.2",
     "react-slide-toggle": "^0.1.13",
     "react-tabs": "^3.0.0",
     "resolve": "1.8.1",
     "sass-loader": "7.1.0",
     "style-loader": "0.23.0",
     "svg-inline-react": "^3.1.0",
-    "sweetalert": "^2.1.2",
     "terser-webpack-plugin": "1.1.0",
     "url-loader": "1.1.1",
     "webpack": "4.19.1",
